# Optimize pod startup times and drastically reduce image sizes