--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -125,6 +125,7 @@ jobs:
     needs: [build-and-push]
     runs-on: ubuntu-latest
     if: github.event_name == 'push' && github.ref == 'refs/heads/master'
+    timeout-minutes: 20
     
     steps:
     - uses: actions/checkout@v4
@@ -155,23 +156,26 @@ jobs:
     - name: Install Playwright browsers
       run: |
         cd frontend
-        npx playwright install --with-deps
+        npx playwright install --with-deps chromium
       if: steps.playwright-cache.outputs.cache-hit != 'true'
     
-    - name: Start services with docker compose
-      run: |
-        docker compose up -d
-        sleep 30
-    
-    - name: Run E2E tests
+    - name: Create test results directory
       run: |
         cd frontend
-        npm run test:e2e
+        mkdir -p test-results
+    
+    - name: Run simplified E2E tests
+      run: |
+        cd frontend
+        npm run test:e2e:ci
+      timeout-minutes: 15
     
     - name: Upload test results
       uses: actions/upload-artifact@v4
       if: always()
       with:
         name: playwright-report
-        path: frontend/playwright-report/
+        path: |
+          frontend/playwright-report/
+          frontend/test-results/
         retention-days: 30