--- a/backend/accounts-service/pom.xml
+++ b/backend/accounts-service/pom.xml
@@ -21,7 +21,7 @@
     <description>Accounts Service for Banking Application</description>
 
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
         <spring-cloud.version>2022.0.4</spring-cloud.version>
         <opentelemetry.version>1.32.0</opentelemetry.version>
     </properties>