--- a/backend/api-gateway/Dockerfile
+++ b/backend/api-gateway/Dockerfile
@@ -14,31 +14,18 @@ COPY src ./src
 RUN mvn clean package -DskipTests
 
 # Runtime image
-FROM --platform=$TARGETPLATFORM amazoncorretto:20-alpine
+FROM --platform=$TARGETPLATFORM gcr.io/distroless/java20-debian12
 
 WORKDIR /app
-
-# Install curl for health checks and create non-root user
-RUN apk add --no-cache curl && \
-    addgroup -g 1001 -S appuser && \
-    adduser -S appuser -u 1001
 
 # Copy the built JAR file
 COPY --from=builder /app/target/api-gateway-*.jar app.jar
 
-# Create logs directory
-RUN mkdir -p /var/log && \
-    chown -R appuser:appuser /var/log /app
-
-# Switch to non-root user
-USER appuser
+# Switch to non-root user (distroless default is nonroot:65534)
+USER 65534
 
 # Expose port
 EXPOSE 80
-
-# Health check
-HEALTHCHECK --interval=30s --timeout=10s --start-period=60s --retries=3 \
-    CMD curl -f http://localhost:80/actuator/health || exit 1
 
 # Run the application with optimized JVM flags for faster startup
 ENTRYPOINT ["java", \
@@ -50,7 +37,6 @@ ENTRYPOINT ["java", \
     "-XX:+UseCompressedOops", \
     "-XX:+UseCompressedClassPointers", \
     "-XX:+TieredCompilation", \
-    "-XX:TieredStopAtLevel=1", \
     "-Djava.security.egd=file:/dev/./urandom", \
     "-Dspring.main.lazy-initialization=true", \
     "-Dspring.main.banner-mode=off", \