… E2E_OPTIMIZATION_SUMMARY.md
… backend/api-gateway/pom.xml
… backend/transactions-service/pom.xml
… backend/user-service/pom.xml
… frontend/e2e/README.md
… frontend/next.config.test.ts
… frontend/package.json
… kubernetes/base/deployments/accounts-service-deployment.yaml
… kubernetes/base/deployments/api-gateway-deployment.yaml
… kubernetes/base/deployments/frontend-deployment.yaml
… kubernetes/base/deployments/transactions-service-deployment.yaml
… kubernetes/base/deployments/user-service-deployment.yaml