Major improvements:
- Reduced health check delays from 120s to 30s for liveness, 60s to 15s for readiness
- Added startup probes for better health monitoring during initialization
- Increased CPU requests from 100m to 200m+ for faster JVM startup
- Changed imagePullPolicy from Always to IfNotPresent to avoid unnecessary pulls

Docker image optimizations:
- Java services: Switched from Amazon Corretto Alpine (~400MB) to Google Distroless (~100MB)
- Frontend: Optimized from 1.8GB to ~50MB using Next.js standalone + distroless
- Removed curl dependencies and unnecessary packages from all images
- Implemented Spring Boot layered JARs for better Docker layer caching
- Removed conflicting JVM flags (-XX:TieredStopAtLevel=1) for better performance

Dependency management:
- Optimized initContainer checks with faster polling (1s vs 2s)
- Streamlined API Gateway service dependency waiting logic

Expected results:
- Pod startup time: 2+ minutes → 30-60 seconds
- Image sizes: Java 400MB → 100MB (75% reduction), Frontend 1.8GB → 50MB (97% reduction)
- Pull times: 5-10x faster
- Build caching: Significantly improved with layered JARs

🤖 Generated with [Claude Code](https://claude.ai/code)